+++ NEWS.md
@@ -1,5 +1,12 @@
 # 📰 PDFFile News
 
+## v0.2.1
+
+- from_datetime method becomes to_pdf_date and coerces dates from string with
+  dateutil to pdf date strings.
+- from_bool method becomes to_xml_bool and coerces strings and other values to
+  xml bool string.
+
 ## v0.2.0
 
 - Automatically converts pdf datestrings to python datetimes and back.